// types.ts
export interface PostProfile {
  full_name: string | null;
  avatar_url: string | null;
}
// ...
export interface RentalPostLike {
  user_id: string;
}

export interface RentalPostComment {
  id: string;
  content: string;
  created_at: string;
  user_id: string;
  post_id: string; // post_id refers to rental_post_id
  profiles: PostProfile | null;
}

export type PaymentTerm = 'monthly' | 'quarterly' | 'semi_annually';

export interface RentalPost {
  id: string;
  created_at: string;
  user_id: string;
  region: string;
  address: string;
  street_name: string;
  room_count: number;
  condition: string;
  image_urls: string[];
  rent_amount: number;
  payment_term: PaymentTerm;
  latitude: number | null;
  longitude: number | null;
  map_link: string | null;

  // Joined data
  profiles: PostProfile | null;
  rental_post_likes: RentalPostLike[];
  rental_post_comments: [{ count: number }];
}